rlvr/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
pydantic>=2.0
numpy>=1.24.0
openai>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
python-dotenv>=1.0.0 